Split off Instruction into its own module
